# docs/swagger.yaml
basePath: /
definitions:
  main.Recipe:
    properties:
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      publishedAt:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - ingredients
    - instructions
    - name
    - tags
    type: object
info:
  contact: {}
  description: Simple REST API for managing recipes
  title: Recipes API
  version: "1.0"
paths:
  /recipes:
    get:
      description: Get all recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Recipe'
            type: array
      summary: List all recipes
      tags:
      - recipes
    post:
      consumes:
      - application/json
      parameters:
      - description: Recipe data
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/main.Recipe'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Recipe'
        "400":
          description: Bad Request
      summary: Create a new recipe
      tags:
      - recipes
  /recipes/{id}:
    delete:
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      summary: Delete a recipe
      tags:
      - recipes
    put:
      consumes:
      - application/json
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      - description: Recipe data
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/main.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Recipe'
        "404":
          description: Not Found
      summary: Update a recipe
      tags:
      - recipes
  /recipes/search:
    get:
      parameters:
      - description: Recipe tag
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Recipe'
            type: array
      summary: Search recipes by tag
      tags:
      - recipes
swagger: "2.0"